wip: particle effects to follower
