+++ Assets/Scripts/Color/PlayEffectOnSwap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3e8dd4f28c22cf4d88122ee4475cada
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
